exercise more #stringToObject and #objectToString
More specifically, make them behave nicely in case of nil.
It should be noted that special care was needed for
 #stringToObject: nil
Indeed it crashes in ORBit call CORBA_ORB_string_to_object(orb, NULL, ev)
because of strncmp direct invocation.

I was not able to find a mention in the spec, may that be [ST1.0] or
[CORBA2.0]. Looking at JacORB behaviour it looks like this return null,
hence mimicking this for Smalltalk. It's also consistent with
 #objectToString: nil default ORBit invocation.