add Number>>#asORBitAny
This is first try for CORBA any. As a side effect of this commit
we dig further into ORBit internal structures; but I think we need
to so that we can find at least the repoId.

What's missing now is the other way around: given an Any convert it
to a Smalltalk object.